# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fashion Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Products ──────────────────────────────────────────────────────────
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{variantId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      parameters:
        - name: variantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      parameters:
        - name: variantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidateInput"
      responses:
        "200":
          description: Valid coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{id}:
    patch:
      operationId: updateCoupon
      tags: [coupons]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Banners ───────────────────────────────────────────────────────────
  /banners:
    get:
      operationId: listBanners
      tags: [banners]
      responses:
        "200":
          description: Banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [banners]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /banners/{id}:
    patch:
      operationId: updateBanner
      tags: [banners]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: deleteBanner
      tags: [banners]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Notifications ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ───────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ── User ───────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatar:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Category:
      type: object
      required: [id, nameEn, nameAr]
      properties:
        id:
          type: integer
        nameEn:
          type: string
        nameAr:
          type: string
        slug:
          type: string
        imageUrl:
          type: ["string", "null"]
        parentCategoryId:
          type: ["integer", "null"]
        productCount:
          type: integer
    CategoryInput:
      type: object
      required: [nameEn, nameAr]
      properties:
        nameEn:
          type: string
        nameAr:
          type: string
        slug:
          type: string
        imageUrl:
          type: string
        parentCategoryId:
          type: ["integer", "null"]
    CategoryUpdate:
      type: object
      properties:
        nameEn:
          type: string
        nameAr:
          type: string
        slug:
          type: string
        imageUrl:
          type: string
        parentCategoryId:
          type: ["integer", "null"]

    # ── Product ────────────────────────────────────────────────────────
...
    CartItem:
      type: object
      required: [variantId, productId, nameEn, nameAr, price, quantity, color, size]
      properties:
        variantId:
          type: integer
        productId:
          type: integer
        nameEn:
          type: string
        nameAr:
          type: string
        imageUrl:
          type: ["string", "null"]
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        quantity:
          type: integer
        color:
          type: string
        size:
          type: string
        stockQuantity:
          type: integer
    Cart:
      type: object
      required: [items, subtotal, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
    CartItemInput:
      type: object
      required: [variantId, quantity]
      properties:
        variantId:
          type: integer
        quantity:
          type: integer
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer

    # ── Review ─────────────────────────────────────────────────────────
...
    Coupon:
      type: object
      required: [id, code, discountType, discountValue, active]
      properties:
        id:
          type: integer
        code:
          type: string
        discountType:
          type: string
        discountValue:
          type: number
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        usageLimit:
          type: ["integer", "null"]
        usageCount:
          type: integer
        active:
          type: boolean
    CouponInput:
      type: object
      required: [code, discountType, discountValue]
      properties:
        code:
          type: string
        discountType:
          type: string
        discountValue:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        usageLimit:
          type: integer
    CouponUpdate:
      type: object
      properties:
        code:
          type: string
        discountType:
          type: string
        discountValue:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        usageLimit:
          type: integer
        active:
          type: boolean
    CouponValidateInput:
      type: object
      required: [code]
      properties:
        code:
          type: string

    # ── Banner ─────────────────────────────────────────────────────────
    Banner:
      type: object
      required: [id, titleEn, titleAr, imageUrl, active, sortOrder]
      properties:
        id:
          type: integer
        titleEn:
          type: string
        titleAr:
          type: string
        subtitleEn:
          type: ["string", "null"]
        subtitleAr:
          type: ["string", "null"]
        imageUrl:
          type: string
        linkUrl:
          type: ["string", "null"]
        active:
          type: boolean
        sortOrder:
          type: integer
    BannerInput:
      type: object
      required: [titleEn, titleAr, imageUrl]
      properties:
        titleEn:
          type: string
        titleAr:
          type: string
        subtitleEn:
          type: string
        subtitleAr:
          type: string
        imageUrl:
          type: string
        linkUrl:
          type: string
        active:
          type: boolean
        sortOrder:
          type: integer
    BannerUpdate:
      type: object
      properties:
        titleEn:
          type: string
        titleAr:
          type: string
        subtitleEn:
          type: string
        subtitleAr:
          type: string
        imageUrl:
          type: string
        linkUrl:
          type: string
        active:
          type: boolean
        sortOrder:
          type: integer

    # ── Notification ───────────────────────────────────────────────────
...